fix(i18n): rename japanese-charset fixture to drop .test suffix
The file was a test data/fixture module (exports only, no test cases)
but vitest picked it up as an empty test suite. Renamed .test.ts → .ts
and updated the import in the validation test.
